Auth pushed to partial class. Test controller renamed and adjusted.
