--- src/CheckBook.App/ApiControllers/TestApiController.cs
@@ -1,15 +1,0 @@
-﻿using System.Web.Http;
-
-namespace CheckBook.App.ApiControllers
-{
-    [AllowAnonymous]
-    public class TestApiController : AppApiController
-    {
-        [Route("is-authenticated")]
-        [HttpGet]
-        public bool IsAuthenticated()
-        {
-            return RequestContext.Principal.Identity.IsAuthenticated;
-        }
-    }
-}
