+++ src/CheckBook.App/Web.config
@@ -10,6 +10,7 @@
     <add key="ida:ClientId" value="" />
     <!--If you are using this in multi-tenant mode, add coma-delimated list of issuers tenant IDs -->
     <add key="ida:Tenants" value="" />
+    <add key="ida:Audiences" value="" />
   </appSettings>
 
   <connectionStrings>
